splits Data class into two - one for access to database and one for importing/exporting
