+++ CocktailApp/Forms/EditCocktailsForm.cs
@@ -17,11 +17,11 @@ namespace CocktailApp.Forms
         {
             InitializeComponent();
 
-            Data.GetIngredients();
+            DataAccess.GetIngredients();
 
             RefreshListContent();
 
-            IngredientTagsComboBox.Items.AddRange(Data.Ingredients.Select(i => i.Type).Distinct().ToArray());
+            IngredientTagsComboBox.Items.AddRange(DataAccess.Ingredients.Select(i => i.Type).Distinct().ToArray());
 
             IngredientTagListBox.DataSource = TagList;
         }
@@ -37,9 +37,9 @@ namespace CocktailApp.Forms
 
         private void SearchButton_Click(object sender, EventArgs e)
         {
-            Data.SearchCocktails(SearchBar.Text, SearchByNameRadioButton.Checked ? "Name" : "Ingredients");
+            DataAccess.SearchCocktails(SearchBar.Text, SearchByNameRadioButton.Checked ? "Name" : "Ingredients");
 
-            CocktailsListBox.DataSource = Data.Cocktails;
+            CocktailsListBox.DataSource = DataAccess.Cocktails;
         }
 
         private void ImportButton_Click(object sender, EventArgs e)
@@ -51,7 +51,7 @@ namespace CocktailApp.Forms
 
             if (dialog.ShowDialog() == DialogResult.OK)
             {
-                Data.ImportCocktails(dialog.FileName);
+                DataTransfer.ImportCocktails(dialog.FileName);
 
                 RefreshListContent();
             }
@@ -63,7 +63,7 @@ namespace CocktailApp.Forms
 
             if (dialog.ShowDialog() == DialogResult.OK)
             {
-                Data.ExportCocktails(dialog.SelectedPath);
+                DataTransfer.ExportCocktails(dialog.SelectedPath);
             }
         }
 
@@ -142,13 +142,13 @@ namespace CocktailApp.Forms
         {
             string name = Format.CapitalizeEvery(NameTextBox.Text);
 
-            if (Data.Cocktails.Select(c => c.Name).Contains(name))
+            if (DataAccess.Cocktails.Select(c => c.Name).Contains(name))
             {
                 SuccessLabelNotUnique();
             }
             else if (TextBoxValidation())
             {
-                Data.AddEditCocktail(
+                DataAccess.AddEditCocktail(
                     name: name,
                     ingredients: string.Join("|", TagList),
                     fullIngredients: string.Join("|", FullIngredientInfoTextBox.Text
@@ -176,7 +176,7 @@ namespace CocktailApp.Forms
             {
                 SuccessLabelNoSelection();
             }
-            else if (Data.Cocktails.Where(c => c.Id.ToString() != id).Select(c => c.Name).Contains(name) && DialogResult.No == 
+            else if (DataAccess.Cocktails.Where(c => c.Id.ToString() != id).Select(c => c.Name).Contains(name) && DialogResult.No == 
                 MessageBox.Show("This will overwrite an existing cocktail with that name.\r\n\r\nDo you want to continue?",
                                 "Edit cocktail", MessageBoxButtons.YesNo))
             {
@@ -185,7 +185,7 @@ namespace CocktailApp.Forms
             }
             else if (TextBoxValidation())
             {
-                Data.AddEditCocktail(
+                DataAccess.AddEditCocktail(
                     id: id,
                     name: name,
                     ingredients: string.Join("|", TagList),
@@ -209,7 +209,7 @@ namespace CocktailApp.Forms
         {
             if (CocktailsListBox.SelectedItems.Count > 0)
             {
-                Data.RemoveCocktail(CocktailsListBox.SelectedValue.ToString());
+                DataAccess.RemoveCocktail(CocktailsListBox.SelectedValue.ToString());
 
                 RefreshListContent();
 
@@ -233,9 +233,9 @@ namespace CocktailApp.Forms
 
         private void RefreshListContent()
         {
-            Data.GetCocktails();
+            DataAccess.GetCocktails();
 
-            CocktailsListBox.DataSource = Data.Cocktails;
+            CocktailsListBox.DataSource = DataAccess.Cocktails;
 
             CocktailsListBox.ClearSelected();
             NameTextBox.Text = "";
